model validation and exceptions

+++ Social/Social.Application/UserProfiles/CommandHandlers/DeleteUserProfileHandler.cs
@@ -1,24 +1,44 @@
 ﻿using MediatR;
 using Microsoft.EntityFrameworkCore;
+using Social.Application.Enums;
+using Social.Application.Models;
 using Social.Application.UserProfiles.Commands;
 using Social.Dal;
+using Social.Domain.Aggregates.UserProfileAggregate;
 
 namespace Social.Application.UserProfiles.CommandHandlers
 {
-    internal class DeleteUserProfileHandler : IRequestHandler<DeleteUserProfile>
+    internal class DeleteUserProfileHandler : IRequestHandler<DeleteUserProfile, OperationResult<UserProfile>>
     {
         private readonly DataContext _ctx;
         public DeleteUserProfileHandler(DataContext ctx)
         {
             _ctx = ctx;
         }
-        public async Task Handle(DeleteUserProfile request, CancellationToken cancellationToken)
+        public async Task<OperationResult<UserProfile>> Handle(DeleteUserProfile request, CancellationToken cancellationToken)
         {
+            var result = new OperationResult<UserProfile>();
+
             var userProfile = await _ctx.UserProfiles
                 .FirstOrDefaultAsync(up => up.Id == request.UserProfileId);
 
+            if (userProfile is null)
+            {
+                result.IsError = true;
+                var error = new Error
+                {
+                    Code = ErrorCode.NotFound,
+                    Message = $"No UserProfile found with ID {request.UserProfileId}"
+                };
+                result.Errors.Add(error);
+                return result;
+            }
+
             _ctx.UserProfiles.Remove(userProfile);
             await _ctx.SaveChangesAsync();
+
+            result.Payload = userProfile;
+            return result;
         }
     }
 }
